\name{bind_response}
\alias{bind_response}
\title{Bind the column with name response...}
\usage{bind_response(d, response.predict, vec)
}
\description{Bind the column with name \code{response.predict} and contents \code{vec} as last column
to data frame d}
\value{dataframe d with column}
\arguments{\item{d}{dataframe}
\item{response.predict}{name of column \code{vec}}
\item{vec}{the last column bound to data frame d}
}

