\name{cat1}
\alias{cat1}
\title{Output the objects to cat if opts$VERBOSE>=1...}
\usage{cat1(opts, ...)
}
\description{Output the objects to \code{cat} if \code{opts$VERBOSE>=1}}
\value{None}
\seealso{\code{\link{cat}}}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

