\name{cat2}
\alias{cat2}
\title{Output the objects to \link{cat} if opts$VERBOSE>=2...}
\usage{cat2(opts, ...)
}
\description{Output the objects to \code{\link{cat}} if \code{opts$VERBOSE>=2}}
\value{None}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

