\name{cma_jTuner}
\alias{cma_jTuner}
\title{Perform CMA-ES tuning.}
\usage{cma_jTuner(confFile, tdm, envT)
}
\description{Perform CMA-ES tuning. Perform a parameter tuning by CMA-ES, using the *Java* 
implementation by Niko Hansen.}
\value{result of CMA-ES algorithm}
\arguments{\item{confFile}{task configuration for tuning algorithm}
\item{tdm}{the TDMR object}
\item{envT}{the environment variable}
}

