% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmReadAndSplit.r
\name{dsetTest}
\alias{dsetTest}
\title{Return test data}
\usage{
dsetTest(x, ...)
}
\arguments{
\item{x}{return value from a prior call to \code{\link{tdmReadAndSplit}}, an object of class \code{\link{TDMdata}}.}

\item{...}{may contain nExp, experiment number}
}
\value{
\code{dset}, a data frame with all train-validation records
}
\description{
Return test data for an object.
See \code{\link{dsetTest.TDMdata}} for details.
}
\seealso{
\code{\link{dsetTest.TDMdata}}   \code{\link{dsetTrnVa.TDMdata}}
}
\keyword{internal}
