\name{dsetTrnVa.TDMdata}
\alias{dsetTrnVa.TDMdata}
\title{Return the train-validation part of a TDMdata object containing the task data.}
\usage{\method{dsetTrnVa}{TDMdata}(x)
}
\description{Return the train-validation part of a \code{TDMdata} object containing the task data.}
\seealso{\code{\link{tdmSplitTestData}}}
\author{Wolfgang Konen, FHK, Feb'2012}
\arguments{\item{x}{return value from a prior call to \code{\link{tdmSplitTestData}}, an object of class \code{TDMdata}.}
}

