% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmDispatchTuner.r
\name{lhdTuner}
\alias{lhdTuner}
\title{Perform LHD tuning.}
\usage{
lhdTuner(confFile, theSpotPath, tdm, envT, dataObj)
}
\arguments{
\item{confFile}{task configuration for tuning algorithm}

\item{theSpotPath}{needed for spot}

\item{tdm}{the TDMR object}

\item{envT}{the environment variable}

\item{dataObj}{the \code{\link{TDMdata}} object containing the data set (train/vali part and test part)}
}
\value{
the result of LHD sampling, i.e. the list \code{spotConfig}
}
\description{
Perform a parameter tuning using a Latin hypercube design (LHD) 
for obtaining best design points. LHD is performed by configuring SPOT 
in such a way that all the budget is used for the initial design (usually LHD).
}
\keyword{internal}
