% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmGeneralUtils.r
\name{Opts}
\alias{Opts}
\alias{Opts.TDMclassifier}
\alias{Opts.TDMenvir}
\alias{Opts.TDMregressor}
\alias{Opts.default}
\alias{Opts.tdmClass}
\alias{Opts.tdmRegre}
\title{Return the list 'opts'.}
\usage{
Opts(x, ...)

\method{Opts}{TDMenvir}(x, ...)

\method{Opts}{TDMclassifier}(x, ...)

\method{Opts}{TDMregressor}(x, ...)

\method{Opts}{tdmClass}(x, ...)

\method{Opts}{tdmRegre}(x, ...)

\method{Opts}{default}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{TDMclassifier}}, \code{\link[=tdmClassify]{tdmClass}}, 
\code{\link{TDMregressor}} or \code{\link[=tdmRegress]{tdmRegre}}.}

\item{...}{-- currently not used --}
}
\value{
the list \code{opts} with DM-specific settings contained in the specified object
}
\description{
Returns the list \code{opts} from objects of class \code{\link{TDMenvir}}, \code{\link{TDMclassifier}},
                 \code{\link{TDMregressor}}, \code{\link[=tdmClassify]{tdmClass}} or \code{\link[=tdmRegress]{tdmRegre}}.
}

