\name{powellTuner}
\alias{powellTuner}
\title{Perform Powell's tuning.}
\usage{powellTuner(confFile, tdm, envT)
}
\description{Perform Powell's tuning. Perform a parameter tuning by Powell's UObyQA algorithm 
(unconstrained optimization by quadratic approximation), see R-package \code{powell}}
\value{result of Powell's algorithm}
\arguments{\item{confFile}{task configuration for tuning algorithm}
\item{tdm}{the TDMR object}
\item{envT}{the environment variable}
}

