\name{print.TDMdata}
\alias{print.TDMdata}
\title{Print an overview for a TDMdata object.}
\usage{\method{print}{TDMdata}(x, ...)
}
\description{Print an overview for a \code{TDMdata} object.}
\details{Print number of rows and number of columns of the data frame \code{dset} contained in the \code{TDMdata} object.}
\seealso{\code{\link{tdmSplitTestData}}}
\author{Wolfgang Konen, FHK}
\arguments{\item{x}{return value from a prior call to \code{\link{tdmSplitTestData}}, an object of class \code{TDMdata}.}
\item{...}{currently not used}
}

