% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printTDMclassifier.r
\name{print.TDMclassifier}
\alias{print.TDMclassifier}
\alias{print.tdmClass}
\title{Print an overview for a \code{\link{TDMclassifier}} or \code{\link[=tdmClassify]{tdmClass}} object.}
\usage{
\method{print}{TDMclassifier}(x, ...)

\method{print}{tdmClass}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[=tdmClassify]{tdmClass}}, as returned from a prior call to \code{\link{tdmClassify}}, \cr
       or  an object of class \code{\link{TDMclassifier}}, as returned from a prior call to \code{\link{tdmClassifyLoop}}.}

\item{...}{e.g. 'type'    which information to print:
     \describe{
     \item{\code{"overview"}}{ (default) relative gain on training/test set, number of records, see \code{\link{tdmClassifySummary}}}
     \item{\code{"cm.train"}}{ confusion matrix on train set}
     \item{\code{"cm.vali"}}{ confusion matrix on test set}
     \item{\code{"?"}}{ help on this method}
     }}
}
\description{
Print an overview for a \code{\link{TDMclassifier}} or \code{\link[=tdmClassify]{tdmClass}} object.


}
\author{
Wolfgang Konen, FHK, 2011 - 2013
}
\seealso{
\code{\link{tdmClassify}}, \code{\link{tdmClassifySummary}}, \code{\link{TDMclassifier}}
}

