\name{print1}
\alias{print1}
\title{Prints the objects using print if opts$VERBOSE>=1...}
\usage{print1(opts, ...)
}
\description{Prints the objects using \code{print} if \code{opts$VERBOSE>=1}}
\value{None}
\seealso{\code{\link{print}}}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

