\name{print2}
\alias{print2}
\title{Prints the objects using \link{print} if opts$VERBOSE>=2...}
\usage{print2(opts, ...)
}
\description{Prints the objects using \code{\link{print}} if \code{opts$VERBOSE>=2}}
\value{None}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

