% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEmbedDataFrame.r
\name{tdmEmbedDataFrame}
\alias{tdmEmbedDataFrame}
\title{Embed columns in a data frame.}
\usage{
tdmEmbedDataFrame(x, cols = NULL, dimension = 1, take = 1)
}
\arguments{
\item{x}{The data frame containing the columns to embed.}

\item{cols}{[NULL] A vector with a list of names of the columns to embed. If NULL, embed all columns of x.}

\item{dimension}{[1] The additional dimensions to generate when embedding.}

\item{take}{[1] Take from these dimensions only every take'th column.}
}
\value{
The data frame, augmented with embedded columns, shortended by \code{dimension} rows.
}
\description{
Embeds the columns named \code{cols} in the data frame \code{x} into a space of dimension \code{dimension}.
}

