\name{tdmEnvTAddGetters}
\alias{tdmEnvTAddGetters}
\title{Add getter functions getBst and getRes to environment envT}
\usage{
  tdmEnvTAddGetters(envT)
}
\arguments{
  \item{envT}{the TDMR environment}
}
\value{
  the augmented \code{envT}
}
\description{
  Add getter functions getBst and getRes to environment
  envT
}

