% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmEnvTGetOpts}
\alias{tdmEnvTGetOpts}
\title{Return list opts from the \code{k}-th \code{.conf}-file}
\usage{
tdmEnvTGetOpts(envT, k = 1)
}
\arguments{
\item{envT}{environment TDMR}

\item{k}{[1] index 1,...,length(envT$runList)}
}
\value{
opts
}
\description{
Return list opts from the \code{k}-th \code{.conf}-file
}

