% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmEnvTLoad}
\alias{tdmEnvTLoad}
\title{Load an \code{envT}-type environment from file \code{fileRData}.}
\usage{
tdmEnvTLoad(fileRData, path = NULL)
}
\arguments{
\item{fileRData}{string with filename to load.}

\item{path}{[NULL] dir where to search \code{fileRData}. If NULL, use current dir.}
}
\value{
envT
}
\description{
The loaded envT is augmented with getter functions, see \code{\link{tdmEnvTAddGetters}}.
}
