% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmEnvTSetOpts}
\alias{tdmEnvTSetOpts}
\title{Set list opts for the \code{k}-th element of \code{envT$sCList}}
\usage{
tdmEnvTSetOpts(envT, opts, k = 1)
}
\arguments{
\item{envT}{environment TDMR}

\item{opts}{list of options}

\item{k}{[1] index 1,...,length(envT$runList)}
}
\value{
envT
}
\description{
Set list opts for the \code{k}-th element of \code{envT$sCList}
}
