% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmEnvTUpdate}
\alias{tdmEnvTUpdate}
\title{Update \code{envT$tdm}}
\usage{
tdmEnvTUpdate(envT, tdm)
}
\arguments{
\item{envT}{environment TDMR}

\item{tdm}{list for TDMR, see \code{\link{tdmDefaultsFill}}}
}
\value{
envT
}
\description{
Update \code{envT$tdm} with the non-NULL elements of \code{tdm}
}
