\name{tdmGraAndLogFinalize}
\alias{tdmGraAndLogFinalize}
\title{Finalize graphics and log file...}
\usage{tdmGraAndLogFinalize(opts)
}
\description{Finalize graphics and log file}
\arguments{\item{opts}{with \code{opts$GD.DEVICE} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
}

