\name{tdmGraAndLogInitialize}
\alias{tdmGraAndLogInitialize}
\title{Initialize graphics and log file.}
\usage{tdmGraAndLogInitialize(opts)
}
\description{Initialize graphics and log file.}
\details{The log file is opened in \code{opts$dir.output/opts$LOGFILE}.}
\arguments{\item{opts}{with \code{opts$GD.DEVICE} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
}

