\name{tdmGraphicCloseDev}
\alias{tdmGraphicCloseDev}
\title{Close all open graphic devices.}
\usage{
  tdmGraphicCloseDev(opts, ...)
}
\arguments{
  \item{opts}{with \code{opts$GD.DEVICE} one out of
  [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}

  \item{...}{optional arguments (currently not used)}
}
\description{
  Close all open graphic devices.
}

