\name{tdmGraphicCloseWin}
\alias{tdmGraphicCloseWin}
\title{Close active file ("png").}
\usage{
  tdmGraphicCloseWin(opts, ...)
}
\arguments{
  \item{opts}{with \code{opts$GD.DEVICE} one out of [
  \code{"pdf"} | \code{"png"} | \code{"win"} |
  \code{"rstudio"} | \code{"non"} ], see
  \code{\link{tdmGraphicInit}}}

  \item{...}{optional arguments (currently not used)}
}
\description{
  Close active file ("png").
}

