\name{tdmGraphicInit}
\alias{tdmGraphicInit}
\title{Initialize graphic device.}
\usage{tdmGraphicInit(opts, ...)
}
\description{Initialize graphic device. Open multipage PDF or (create and) clear opts$GD.PNGDIR.}
\arguments{\item{opts}{with \code{opts$GRAPHDEV} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
\item{...}{optional arguments to hand over to \code{\link{pdf}}}
}

