% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmGraphicUtils.r
\name{tdmGraphicToTop}
\alias{tdmGraphicToTop}
\title{Bring the acitve window to the top}
\usage{
tdmGraphicToTop(opts)
}
\arguments{
\item{opts}{with \code{opts$GD.DEVICE} one out of 
[ \code{"pdf"} | \code{"png"} | \code{"win"} | \code{"rstudio"} | \code{"non"} ],
see \code{\link{tdmGraphicInit}}}
}
\description{
Only relevant for opts$GD.DEVICE=="win".
}
