\docType{data}
\name{tdmMapDesSpot}
\alias{tdmMapDesSpot}
\title{Helper fct for \code{\link{tdmStartSpot}}.}
\format{List of 2
 $ load :function (tdm)  
 $ apply:function (des, opts, k, tdm, spotConfig)}
\usage{
  tdmMapDesSpot
}
\description{
  Map the parameters from \code{des} to \code{opts} for the
  \code{k}th line of \code{des}.  The object tdmMapDesSpot
  has two function elements: \cr \code{load(tdm)} \cr
  \code{apply(des,opts,k,tdm)}
}
\keyword{datasets}
\keyword{internal}

