% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmModelingUtils.r
\name{tdmModVote2Target}
\alias{tdmModVote2Target}
\title{Analyze how the vote fraction corresponds to reliability of prediction.}
\usage{
tdmModVote2Target(vote0, pred, target)
}
\arguments{
\item{vote0}{vector: which fraction of trees votes for class 0?}

\item{pred}{vector: the predicted class for each record (0/1)}

\item{target}{vector: the true class for each vector (0/1)}
}
\value{
a data frame with columns 
   \item{vcut}{  vote cut v}
   \item{count}{ number of cases with vote fraction in [v[i-1],v[i]]}
   \item{pred0}{ fraction of 0-predictions}
   \item{pCorr}{ fraction of correct predictions}
   \item{pR}{    fraction of true 1-cases}
}
\description{
This function analyzes whether in different vote bins the trained RF makes
predictions with different reliability. Only for RF-prediction in case of 
binary (0/1) classification. \cr\cr Expected result: The larger the fraction
of trees voting for class 0 is, the smaller is the percentage of true class-1-
cases in this vote bin.
This function is somewhat specialized for the DMC2010-task.
}
\author{
Wolfgang Konen \email{wolfgang.konen@th-koeln.de}
}

