\name{tdmPreFindConstVar}
\alias{tdmPreFindConstVar}
\title{Find constant columns.}
\usage{tdmPreFindConstVar(dset)
}
\description{Find constant columns. Find all those columns in data frame \code{dset} which are completely constant or
completely NA and return a vector with their names.}
\value{name vector of constant columns}
\arguments{\item{dset}{data frame}
}

