% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmReadAndSplit.r
\name{tdmReadTaskData}
\alias{tdmReadTaskData}
\title{Read task data.}
\usage{
tdmReadTaskData(envT, tdm)
}
\arguments{
\item{envT}{environment TDMR}

\item{tdm}{list with TDMR controls}
}
\value{
\code{dataObj}, see \code{\link{tdmReadAndSplit}}
}
\description{
Read and split task data (wrapper for \code{\link{tdmReadAndSplit}}).
}
\seealso{
\code{\link{dsetTrnVa.TDMdata}}, \code{\link{dsetTest.TDMdata}}, \code{\link{tdmReadAndSplit}}
}
