% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmReportSensY}
\alias{tdmReportSensY}
\title{Sensitivity Report Helper Function}
\usage{
tdmReportSensY(B, fit, roi, nsens)
}
\arguments{
\item{B}{best solution column}

\item{fit}{random forest fit}

\item{roi}{internal parameter for the initial region of interest, or: \code{spotConfig$alg.roi}}

\item{nsens}{number of parameters estimated with sensitivity}
}
\value{
Y
}
\description{
Helper function for the sensitivity report
}
\seealso{
\code{\link{tdmEnvTReport}}  \code{\link{tdmEnvTReportSens}}
}
\keyword{internal}
