\name{tdmROCR}
\alias{tdmROCR}
\title{Interactive plot of ROC, lift or other charts for a \code{\link{TDMclassifier}} object.
  See \code{\link{tdmROCR.TDMclassifier}} for details}
\usage{
  tdmROCR(x, ...)
}
\arguments{
  \item{x}{return value from a prior call to
  \code{\link{tdmClassifyLoop}}, an object of class
  \code{\link{TDMclassifier}}.}

  \item{...}{-- currently not used --}
}
\description{
  Interactive plot of ROC, lift or other charts for a
  \code{\link{TDMclassifier}} object.  See
  \code{\link{tdmROCR.TDMclassifier}} for details
}
\seealso{
  \code{\link{tdmROCR.TDMclassifier}}
  \code{\link{tdmROCRbase}}
}
\keyword{internal}

