% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmStartSpot2.r
\name{tdmStartSpot2}
\alias{tdmStartSpot2}
\title{Function called by SPOT 2.0 tuner}
\usage{
tdmStartSpot2(x, tdm, envT, dataObj, opts = NULL)
}
\arguments{
\item{x}{a matrix with design points (one point per row)}
}
\value{
\code{y}, a 1-column matrix, same number of rows as \code{x} 
    returns the fitness and saves other
    diagnostic results in envT$spotConfig, envT$res, envT$bst
}
\description{
This function is called when \code{tdm$tuneMethod} = "spot" or "lhd".
}
\author{
Wolfgang Konen, THK, 2018
}
\keyword{internal}
