% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiasedRun.r
\name{zzz}
\alias{zzz}
\title{just to make example check happy}
\usage{
zzz()
}
\value{
\code{dummy}     zero
}
\description{
If we do not have this little dummy function with the plot command in the example, 
then 'R CMD check' in buildTDMR.r will fail when the examples are checked. 
The reason: The file TDMR.Rcheck/TDMR-Ex.R contains a footer which tries to 
close a graphics device with grDevices::dev.off(). But if there is no graphics device
from the previous example, dev.off() tries to operate on the null device and this
produces an error.
}
\examples{
   plot(1:10);
   

}
\author{
Wolfgang Konen, FHK, 2010 - 2013
}
\keyword{internal}

