\name{cat1}
\alias{cat1}
\title{Output the objects to \link{cat} if opts$VERBOSE>=1...}
\usage{cat1(opts, ...)
}
\description{Output the objects to \code{\link{cat}} if \code{opts$VERBOSE>=1}}
\value{None}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

