\name{cat2}
\alias{cat2}
\title{Output objects to cat if opts$VERBOSE>=2.}
\usage{cat2(opts, ...)
}
\description{Output objects to \code{cat} if \code{opts$VERBOSE>=2}.}
\value{None}
\seealso{\code{\link{cat}}}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

