\name{cat2}
\alias{cat2}
\title{Output objects to \code{cat} if \code{opts$VERBOSE>=2}.}
\usage{
  cat2(opts, ...)
}
\arguments{
  \item{opts}{from which we need the element VERBOSE}

  \item{...}{objects}
}
\value{
  None
}
\description{
  Output objects to \code{cat} if \code{opts$VERBOSE>=2}.
}
\seealso{
  \code{\link{cat}}
}
\keyword{internal}

