% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmDispatchTuner.r
\name{cmaesTuner}
\alias{cmaesTuner}
\title{Perform CMA-ES tuning (R version).}
\usage{
cmaesTuner(confFile, tdm, envT, dataObj)
}
\arguments{
\item{confFile}{task configuration for tuning algorithm}

\item{tdm}{the TDMR object}

\item{envT}{the environment variable}

\item{dataObj}{the \code{\link{TDMdata}} object containing the data set (train/vali part and test part)}
}
\value{
the result of CMA-ES tuning, i.e. the list \code{envT$spotConfig}, extended by
   \item{\code{cma}}{ the return value from \code{\link[cmaes]{cma_es}}   }
   \item{\code{cma$count}}{ the number of calls to \code{tdmStartOther}  }
}
\description{
Perform a parameter tuning by CMA-ES, using the *R*-implementation 
(package \code{cmaes} by Olaf Mersmann).
}
\note{
This tuner \code{cmaesTuner} is deprecated. Use instead \code{\link{cma_jTuner}}.
}
\keyword{internal}
