% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmSplitTestData.r
\name{dsetTest.TDMdata}
\alias{dsetTest.TDMdata}
\title{Return test data of \code{\link{TDMdata}} object}
\usage{
\method{dsetTest}{TDMdata}(x, ...)
}
\arguments{
\item{x}{return value from a prior call to \code{\link{tdmSplitTestData}}, an object of class \code{\link{TDMdata}}.}

\item{...}{may contain nExp, experiment number, needed only if \code{x$tdm$umode=="SP_T"}: 
add nExp to seed when randomly splitting in train and test data [default: nExp=0]}
}
\value{
\code{tset}, a data frame with all test records. If there are 0 test records, return NULL.
}
\description{
Return the test part of a \code{\link{TDMdata}} object containing the task data.
}
\author{
Wolfgang Konen, THK
}
\seealso{
\code{\link{unbiasedRun}} \code{\link{dsetTrnVa.TDMdata}} \code{\link{tdmSplitTestData}}
}

