% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tdmDispatchTuner.r
\name{powellTuner}
\alias{powellTuner}
\title{Perform Powell's tuning.}
\usage{
powellTuner(confFile, tdm, envT, dataObj)
}
\arguments{
\item{confFile}{task configuration for tuning algorithm}

\item{tdm}{the TDMR object}

\item{envT}{the environment variable}

\item{dataObj}{the \code{\link{TDMdata}} object containing the data set (train/vali part and test part)}
}
\value{
the result of Powell tuning, i.e. the list \code{envT$spotConfig}, extended by
   \item{\code{powell}}{ the return value from \code{\link[powell]{powell}}   }
   \item{\code{powell$count}}{ the number of calls to \code{tdmStartOther}  }
}
\description{
Perform a parameter tuning by Powell's UObyQA algorithm
(unconstrained optimization by quadratic approximation), see R-package \code{powell}
}
\keyword{internal}

