\name{print1}
\alias{print1}
\title{Prints the objects using \link{print} if opts$VERBOSE>=1...}
\usage{print1(opts, ...)
}
\description{Prints the objects using \code{\link{print}} if \code{opts$VERBOSE>=1}}
\value{None}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

