% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmGeneralUtils.r
\name{print1}
\alias{print1}
\title{Print objects using \code{print} if \code{opts$VERBOSE>=1}.}
\usage{
print1(opts, ...)
}
\arguments{
\item{opts}{from which we need the element VERBOSE}

\item{...}{objects}
}
\value{
None
}
\description{
Print objects using \code{print} if \code{opts$VERBOSE>=1}.
}
\seealso{
\code{\link{print}}
}
\keyword{internal}

