\name{print2}
\alias{print2}
\title{Print objects using print if opts$VERBOSE>=2.}
\usage{print2(opts, ...)
}
\description{Print objects using \code{print} if \code{opts$VERBOSE>=2}.}
\value{None}
\seealso{\code{\link{print}}}
\arguments{\item{opts}{from which we need the element VERBOSE}
\item{...}{objects}
}

