\name{print2}
\alias{print2}
\title{Print objects using \code{print} if \code{opts$VERBOSE>=2}.}
\usage{
  print2(opts, ...)
}
\arguments{
  \item{opts}{from which we need the element VERBOSE}

  \item{...}{objects}
}
\value{
  None
}
\description{
  Print objects using \code{print} if
  \code{opts$VERBOSE>=2}.
}
\seealso{
  \code{\link{print}}
}
\keyword{internal}

