% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tdmDispatchTuner.r
\name{spotTuner}
\alias{spotTuner}
\title{Perform a \link{SPOT} tuning.}
\usage{
spotTuner(confFile, spotStep, theSpotPath, tdm, envT, dataObj)
}
\arguments{
\item{confFile}{task configuration for tuning algorithm}

\item{spotStep}{needed for spot}

\item{theSpotPath}{needed for spot}

\item{tdm}{the TDMR object}

\item{envT}{the environment variable}

\item{dataObj}{the \code{\link{TDMdata}} object containing the data set (train/vali part and test part)}
}
\value{
the result of SPOT tuning, i.e. the list \code{spotConfig}
}
\description{
\code{spotTuner} calls \code{\link{spot}} and write the necessary information on environment \code{envT}.
}
\keyword{internal}

