% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tdmEnvTMakeNew.r
\name{tdmEnvTLoad}
\alias{tdmEnvTLoad}
\title{Load an \code{envT}-type environment from file \code{fileRData}.}
\usage{
tdmEnvTLoad(fileRData)
}
\arguments{
\item{fileRData}{string with filename to load. This file is searched in the working dir.}
}
\value{
envT
}
\description{
The loaded envT is augmented with getter functions, see \code{\link{tdmEnvTAddGetters}}.
}

