% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tdmGraphicUtils.r
\name{tdmGraAndLogFinalize}
\alias{tdmGraAndLogFinalize}
\title{Finalize graphics and log file}
\usage{
tdmGraAndLogFinalize(opts, gdObj = NULL)
}
\arguments{
\item{opts}{with \code{opts$GD.DEVICE} one out of
             [ \code{"pdf"} | \code{"png"} | \code{"win"} | \code{"rstudio"} | \code{"non"} ],
             see \code{\link{tdmGraphicInit}}}

\item{gdObj}{object of class TDMgdev, the return value from tdmGraAndLogInitialize, to ensure that tdmGraAndLogInitialize was called before
(and the sink on opts$LOGFILE can be closed)}
}
\description{
Finalize graphics and log file
}

