\name{tdmGraAndLogInitialize}
\alias{tdmGraAndLogInitialize}
\title{Initialize graphics and log file...}
\usage{tdmGraAndLogInitialize(opts)
}
\description{Initialize graphics and log file}
\arguments{\item{opts}{with \code{opts$GRAPHDEV} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
}

