\name{tdmGraphicCloseDev}
\alias{tdmGraphicCloseDev}
\title{Close all open graphic devices.}
\usage{tdmGraphicCloseDev(opts, ...)
}
\description{Close all open graphic devices.}
\arguments{\item{opts}{with \code{opts$GD.DEVICE} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
\item{...}{optional arguments (currently not used)}
}

