\name{tdmGraphicCloseWin}
\alias{tdmGraphicCloseWin}
\title{Close active file ("png").}
\usage{tdmGraphicCloseWin(opts, ...)
}
\description{Close active file ("png").}
\arguments{\item{opts}{with \code{opts$GRAPHDEV} one out of [\code{"pdf"},\code{"png"},\code{"win"},\code{"non"}]}
\item{...}{optional arguments (currently not used)}
}

