% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmGraphicUtils.r
\name{tdmGraphicNewWin}
\alias{tdmGraphicNewWin}
\title{Initialize a new window.}
\usage{
tdmGraphicNewWin(opts, ...)
}
\arguments{
\item{opts}{with \code{opts$GD.DEVICE} one out of 
[ \code{"pdf"} | \code{"png"} | \code{"win"} | \code{"rstudio"} | \code{"non"} ],
see \code{\link{tdmGraphicInit}}}

\item{...}{optional arguments to hand over to \code{\link{png}} or \code{windows} or \code{X11} in 
package grDevices (the other devices require no further arguments)}
}
\description{
Initialize a new window ("win") / a new file ("png") for current graphic device.
}
