% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmMapDesign.r
\name{tdmMapDesLoad}
\alias{tdmMapDesLoad}
\title{Load the mapping files.}
\usage{
tdmMapDesLoad(tdm = list())
}
\arguments{
\item{tdm}{list, needed for \code{tdm$tdmPath} and \code{tdm$path}}
}
\value{
\code{tdm}, the modified list with new elements \code{tdm$map} and \code{tdm$mapUser}
}
\description{
Load the map files \code{"tdmMapDesign.csv"} and optionally 
   also \code{"userMapDesign.csv"} and store them in \code{tdm$map} and \code{tdm$mapUser},
   resp. These maps are used by \code{\link{tdmMapDesApply}}.   \cr
   \code{"tdmMapDesign.csv"} is searched in the TDMR library path \code{find.package("TDMR")}.
   (For the developer version: \code{<tdm$tdmPath>/inst}). \cr
   \code{"userMapDesign.csv"} is searched in tdm$path (which is getwd() if the user did not define tdm$path).
}
\seealso{
\code{\link{tdmMapDesApply}}
}
