% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmPreprocUtils.r
\name{tdmPreFindConstVar}
\alias{tdmPreFindConstVar}
\title{Find constant columns.}
\usage{
tdmPreFindConstVar(dset)
}
\arguments{
\item{dset}{data frame}
}
\value{
name vector of constant columns
}
\description{
Find all those columns in data frame \code{dset} which are completely constant or
completely NA and return a vector with their names.
}
